Restructured LaTeX tests, files too large.
